requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
faster-whisper==0.10.0
numpy==1.24.3
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
ollama==0.1.7
langchain==0.1.0
langchain-core==0.1.0
langchain-ollama==0.1.0
httpx==0.25.2
python-dotenv==1.0.0
audioop-lts==0.2.1 